VariLC: Fix VC++ warning
Signed/unsigned mismatch.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@14388 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/VariLC/VariLC.cpp
@@ -700,7 +700,7 @@ int VariLC::OnSendToVariLC(MM::PropertyBase* pProp, MM::ActionType eAct)
 	  else if ((std::string)state=="State") {   
 		  	  
 			  std::vector<char> val(len-5);
-			  for (int i=5; i < len; i++) {				  
+			  for (size_t i=5; i < len; i++) {
 					val[5-i] = sendToVariLC_[i];
 			  }
 			  val[len] = '\0';
